mise en float des pilotes girouette pour tester avec le servo

+++ applications/girouette.h
@@ -5,6 +5,6 @@
 void Init_Girouette(void);
 
 //Retourne l'angle de al girouette(pour le PWM du servomoteur)
-int Return_Angle_Girouette(void);
+float Return_Angle_Girouette(void);
 
 #endif //_GIROUETTE_H__//
